moving getGithubRepoToLoad to file
